allDjangoProjects/rentingHouse/requirements.txt
alabaster==0.7.12
astroid==2.2.5
Babel==2.6.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-reversion==3.0.3
DjangoUeditor==1.8.143
docopt==0.6.2
et-xmlfile==1.0.1
fake-useragent==0.1.11
future==0.17.1
httplib2==0.12.3
idna==2.8
jdcal==1.4
lazy-object-proxy==1.4.0
odfpy==1.4.0
openpyxl==2.6.2
Pillow==5.4.1
pipreqs==0.4.9
PyMySQL==0.9.3
pytz==2018.7
PyYAML==5.1
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.3.5
urllib3==1.24.1
webencodings==0.5.1
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
